+++ ft_printchar.c
@@ -1,10 +1,10 @@
 #include "ft_printf.h"
 
-long int	ft_printchar(p_list *list, va_list arg)
+int		ft_printchar(p_list *list, va_list arg)
 {
-	long int 		len;
+	int 			len;
 	unsigned char	str;
-	long int 		i;
+	int 			i;
 
 	i = 0;
 	str = (char)va_arg(arg, int);
